Separate all Naive device functions.
